[mod] handlerをapp.jsにまとめる
