+++ todos/tests/unit/test-get-handler.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const app = require('../../handlers/getTodo.js');
+const app = require('../../app.js');
 const chai = require('chai');
 const dynamoStub = require("../libs/dynamo-stub")
 const expect = chai.expect;
@@ -14,7 +14,7 @@ describe('Test Index', function () {
 
     it('GetAllData successful getting all data', async () => {
         event = {queryStringParameters: {} }
-        const result = await app.lambdaHandler(event, context)
+        const result = await app.getAllHandler(event, context)
         expect(result).to.be.an('object');
         expect(result.statusCode).to.equal(200);
         expect(result.body).to.be.an('string');
@@ -27,7 +27,7 @@ describe('Test Index', function () {
 
     it('GetAllData successful title search', async () => {
         event = { queryStringParameters: { title: 'test todo'} }
-        const result = await app.lambdaHandler(event, context)
+        const result = await app.getAllHandler(event, context)
         expect(result).to.be.an('object');
         expect(result.statusCode).to.equal(200);
         expect(result.body).to.be.an('string');
@@ -40,7 +40,7 @@ describe('Test Index', function () {
 
     it('GetAllData successful content search', async () => {
         event = { queryStringParameters: { content: 'search test'} }
-        const result = await app.lambdaHandler(event, context)
+        const result = await app.getAllHandler(event, context)
         expect(result).to.be.an('object');
         expect(result.statusCode).to.equal(200);
         expect(result.body).to.be.an('string');
@@ -53,7 +53,7 @@ describe('Test Index', function () {
 
     it('GetAllData successful invalid parameter', async () => {
         event = { queryStringParameters: { test: 'test code request'} }
-        const result = await app.lambdaHandler(event, context)
+        const result = await app.getAllHandler(event, context)
         expect(result).to.be.an('object');
         expect(result.statusCode).to.equal(200);
         expect(result.body).to.be.an('string');
@@ -66,7 +66,7 @@ describe('Test Index', function () {
 
     it('GetAllData successful no result', async () => {
         event = { queryStringParameters: { title: 'nothing result'} }
-        const result = await app.lambdaHandler(event, context)
+        const result = await app.getAllHandler(event, context)
         expect(result).to.be.an('object');
         expect(result.statusCode).to.equal(200);
         expect(result.body).to.be.an('string');
@@ -80,7 +80,7 @@ describe('Test Index', function () {
     it('findData successful', async () => {
         const todoId = 1
         event = { pathParameters: {todo_id: todoId} }
-        const result = await app.findTodoHandler(event, context)
+        const result = await app.findHandler(event, context)
         expect(result).to.be.an('object');
         expect(result.statusCode).to.equal(200);
         expect(result.body).to.be.an('string');
@@ -94,7 +94,7 @@ describe('Test Index', function () {
     it('findData not found', async () => {
         const todoId = 9999
         event = { pathParameters: {todo_id: todoId} }
-        const result = await app.findTodoHandler(event, context)
+        const result = await app.findHandler(event, context)
         expect(result).to.be.an('object');
         expect(result.statusCode).to.equal(404);
         expect(result.body).to.be.an('string');
